+++ Samples/macOS-Swift/macOS-Swift/AppDelegate.swift
@@ -14,7 +14,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
 
     func applicationDidFinishLaunching(_ aNotification: Notification) {
         // Insert code here to initialize your application
-        SentrySDK.with(options: [
+        SentrySDK.start(options: [
             "dsn": "https://8ee5199a90354faf995292b15c196d48@o19635.ingest.sentry.io/4394",
             "debug": true,
             "logLevel": "verbose",