+++ Sources/Sentry/SentrySDK.m
@@ -42,27 +42,27 @@ static SentryHub *currentHub;
 }
  
 + (id)initWithOptions:(NSDictionary<NSString *,id> *)optionsDict {
-    [SentrySDK withOptions:optionsDict];
+    [SentrySDK startWithOptions:optionsDict];
     return nil;
 }
 
 + (id)initWithOptionsObject:(SentryOptions *)options {
-    [SentrySDK withOptionsObject:options];
+    [SentrySDK startWithOptionsObject:options];
     return nil;
 }
 
-+ (void)withOptions:(NSDictionary<NSString *,id> *)optionsDict {
++ (void)startWithOptions:(NSDictionary<NSString *,id> *)optionsDict {
     NSError *error = nil;
     SentryOptions *options = [[SentryOptions alloc] initWithDict:optionsDict didFailWithError:&error];
     if (nil != error) {
         [SentryLog logWithMessage:@"Error while initializing the SDK" andLevel:kSentryLogLevelError];
         [SentryLog logWithMessage:[NSString stringWithFormat:@"%@", error] andLevel:kSentryLogLevelError];
     } else {
-        [SentrySDK withOptionsObject:options];
+        [SentrySDK startWithOptionsObject:options];
     }
 }
 
-+ (void)withOptionsObject:(SentryOptions *)options {
++ (void)startWithOptionsObject:(SentryOptions *)options {
     SentryClient *newClient = [[SentryClient alloc] initWithOptions:options];
     // The Hub needs to be initialized with a client so that closing a session can happen.
     [SentrySDK setCurrentHub:[[SentryHub alloc] initWithClient:newClient andScope:nil]];